+++ src/com/zones/selection/CUIHandler.java
@@ -50,7 +50,7 @@ public class CUIHandler {
         }
     }
     
-    private void sendCUIPoint(int index, ZoneVertice point,int height, int size) {
+    private void sendCUIPoint(int index, ZoneVertice point,int height, long size) {
         getPlayer().sendRawMessage(CUI + join("|","p",index,point.getX(),height,point.getY(),size));
     }
